// ui/src/lib/api-client/models/GetAdminClientRegistrationSettings200Response.ts
export interface GetAdminClientRegistrationSettings200Response {
    /**
     * Whether dynamic client registration is enabled
     * @type {boolean}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    enabled: boolean;
    /**
     * Whether HTTPS is required for redirect URIs
     * @type {boolean}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    requireHttps: boolean;
    /**
     * 
     * @type {Array<string>}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    allowedScopes: Array<string>;
    /**
     * Maximum client lifetime in days (0 = no limit)
     * @type {number}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    maxClientLifetime: number;
    /**
     * Whether terms of service URI is required
     * @type {boolean}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    requireTermsOfService: boolean;
    /**
     * Whether privacy policy URI is required
     * @type {boolean}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    requirePrivacyPolicy: boolean;
    /**
     * Whether public clients are allowed
     * @type {boolean}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    allowPublicClients: boolean;
    /**
     * Whether confidential clients are allowed
     * @type {boolean}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    allowConfidentialClients: boolean;
    /**
     * Whether backend service clients are allowed
     * @type {boolean}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    allowBackendServices: boolean;
    /**
     * Whether admin approval is required for new clients
     * @type {boolean}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    adminApprovalRequired: boolean;
    /**
     * Rate limit for registration requests per minute
     * @type {number}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    rateLimitPerMinute: number;
    /**
     * Maximum number of redirect URIs per client
     * @type {number}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    maxRedirectUris: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    allowedRedirectUriPatterns: Array<string>;
    /**
     * Email to notify of new registrations
     * @type {string}
     * @memberof GetAdminClientRegistrationSettings200Response
     */
    notificationEmail?: string;
}